[Justice Counts] Only Update num_new_datapoints if a non-null datapoint is created (Recidiviz/recidiviz-data#39239)
## Description of the change

Adjusted the calculation of `num_new_datapoints` to only count new
datapoints that have a value. Null datapoints are created when an agency
uploads a breakdown sheet but either doesn't have all breakdowns enabled
or omits some breakdowns from the sheet. If a breakdown is defined in
the metric but missing from the sheet, it gets a null datapoint. With
this change, we only include it in the count if it has a value.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e9f2226787e185957729635023d10ac30a307024