[Raw data] Do not count raw data descriptions starting with TODO as documentation (Recidiviz/recidiviz-data#34096)
## Description of the change

We have a rule that you can't use a raw data column in an ingest view or
downstream BQ view if both the description for that column and the
file-level description are non-empty. This is supposed to deter us from
using data without writing down what we've learned about that data such
that we know how to use it. However, we've gotten around this rule by
filling descriptions with repeated TODOs.

This change stops counting these types of docstrings as "meaningful"
documentation for the purposes of being able to use the data downstream,
with a series of exemptions for existing undocumented columns that are
being used.

I've added three tiers of column documentation exemptions:
* State-level exemptions - All TODO-type columns in this state are still
treated as "documented"
* File-level exemptions - All TODO-type columns in this single raw data
file are treated as "documented"
* Column-level exemptions - These specific columns are treated as
"documented", even if they have TODO-type documentation.

Additionally, I've added exemptions for files that have TODO-type file
descriptions.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#39240
Part of Recidiviz/recidiviz-data#39269

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: edfe113b512f7c30bf0fd8f1a4624537b05af38c