[US_TX] Remove '0001-01-01' start dates from supervision periods (Recidiviz/recidiviz-data#39339)
## Description of the change

This PR removes '0001-01-01' dates from supervision periods. This is
only a temporary fix so that the view graph
[error](https://941882c9ef884317b481f986ea6d3a40-dot-us-central1.composer.googleusercontent.com/log?dag_id=recidiviz-staging_calculation_dag&task_id=update_managed_views_all&execution_date=2025-03-07T23%3A35%3A56.256938%2B00%3A00)
can be resolved. There is an ongoing effort already to refactor
supervision periods in Texas so this will be fully addressed in the
future.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7be8c2ae51ed1ec54001f3f5bba1be0ac2ff8df2