[US_ND] Fix bugs in state staff views (Recidiviz/recidiviz-data#39082)
## Description of the change

This fixes two bugs in two `state_staff` views in ND that were causing
some periods to stay open forever

1. `docstars_staff_location_periods`: this PR takes away the logic in
the view that attempts to discard rows where an officer's location did
not change, since the query only returns <700 rows without that
filtering. It relocates other logic to avoid dropping employment status
updates before they are factored into the final periods.

2. `docstars_staff_role_periods`: I was mistakenly comparing the date of
an update to the date an officer last appeared as active in the data
instead of the most recent time we received data. The goal of that piece
of logic was to close periods when an officer stopped appearing in the
data, so we need to compare the date of the row to the most recent time
we received data, not the most recent time the officer appeared. This
change makes the `role_period` view match the `caseload_type_period`
view.

I confirmed that this has the intended affect on the few officers we
have identified as having an open location or role period and no active
employment at DOCR.

[test
pipeline](https://console.cloud.google.com/dataflow/jobs/us-east1/2025-03-07_10_18_18-1017624055360330346;graphView=0?project=recidiviz-staging&inv=1&invt=Abrakw&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22)))
## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 3a9256897afee2f30f42d548c440dec536dd8e51