[US_UT] Temporarily exclude stable counts validations  (Recidiviz/recidiviz-data#38711)
## Description of the change

UT is not yet sending data regularly, so stable counts violations cannot
provide useful information. This excludes them from running temporarily
to reduce noise, and removes piecemeal exclusions of specific months
that were already put in place. This will make it easier to turn them on
when we are ready.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Related to Recidiviz/recidiviz-data#38599
Related to Recidiviz/recidiviz-data#37244
Related to Recidiviz/recidiviz-data#37245
Related to Recidiviz/recidiviz-data#38405

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 6159a92b3361b8e8dd54ca62d73b13ec4c44180d