[US_AZ][US_ND] Validation adjustments (Recidiviz/recidiviz-data#39412)
## Description of the change

This excludes February 2025 from the
`state_supervision_violation_response_by_response_date_stable_counts`
validation for ND, because there are few enough violations in a standard
month that the short month including a long weekend was enough to make
the validation fail.

It also bumps the failure threshold for
`outliers_staff_count_percent_change_intramonth` to 11%, which is the
rough equivalent of a 2-person difference since the staff is so small.

It also adjusts the standard variation in count of supervision periods
in AZ.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Related to Recidiviz/recidiviz-data#39300
Related to Recidiviz/recidiviz-data#39303
Related to Recidiviz/recidiviz-data#39304
Related to Recidiviz/recidiviz-data#39209

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 0b15608a889c69c66e8287130e7e8cb76cd275d5