[US_IX][Workflows] Replace time served for DUI override to custodial authority sessions (Recidiviz/recidiviz-data#39418)
## Description of the change

Tanya let us know that one of her clients shouldn't be eligible for SLD
because they were serving for a DUI and hadn't spent a full year on
supervision yet. The issue was that we were counting XCRC as time spent
on supervision. Similar to Recidiviz/recidiviz-data#38051, I update the timer served section to
pull from `custodial_authority_sessions`.

Only one client is now ineligible due to this change. (The client in
mention is actually newly eligible as of 3/4/25).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7d049ea8c0baca62b64f2bf7cf6c6494d45d2243