[Deploy] Move post-deploy tasks after git tag creation (Recidiviz/recidiviz-data#39447)
## Description of the change

We got into a weird state today where the staging deploy completed and
the whole calc DAG ran before the on-call had responded `y` to the
post-deploy tasks prompt
([thread](https://recidiviz.slack.com/archives/C028X32LRH7/p1741738905165319)).
It turns out that for staging deploys, we only create / push the new
version tag once that prompt is completed, so we got ourselves in a
weird state where the deployed version of views as recorded in BQ was
ahead of any tag that existed in git.

This PR:
* Moves the git tag creation to happen before the deploy bot
notification and before the post-deploy checklist prompt
* Skips the pre/post-deploy checklist prompts for no-promote (CP staging
release candidate) deploys. This is something @colincadams and I
discussed briefly this week.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

See: https://recidiviz.slack.com/archives/C028X32LRH7/p1741738905165319

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: fa8b1dbb66fd68d7a6cda1cb0096089dc2e5c470