Refactors custom metrics query templating function and deprecates legacy infra (Recidiviz/recidiviz-data#39403)
## Description of the change

Rips out all remaining `legacy` aggregated metrics code and refactors
`get_custom_metrics` to use an updated query templating function that
uses the same perf optimized logic. The structure of the query template
closely mimics that of the looker view, in that there are cte's included
for all units of observation assignments and metric cte's, which all get
combined at the end in one big join. This change shouldn't alter the way
this function is called in notebooks, and my validation so far shows the
same output (with some slight changes to be expected similar to those
surfaced via the looker refactor due to small bug fixes).

Merging this is a pre-req to merging
https://github.com/Recidiviz/recidiviz-data/pull/38807

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#35909

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 51dc50652c90d7584575ccc731a7eaeda78abab3