[NC] Add SFTP user; setup GCS bucket and replication (Recidiviz/recidiviz-data#39406)
## Description of the change

This builds on the terraform configuration that @ohaibbq started and
makes it feature complete for NC. This makes the following changes:

* Adds the SFTP user that they will login as.
* Sets up a specific Kubernetes SA for this pod, that is granted
permissions to write to the GCS bucket. This is the SA that will be used
for any actions taken by the NC sftp user.
* Sets up the GCS bucket in the north carolina specific project, to be
used as the storage for the SFTP server.
* Sets up the replication rule to replicate data from that bucket to
staging and prod automatically (and all permissions to make that work.

Additionally, I tried to set this up in a way such that we can reuse it
for non-sftp states that need a bucket in their own project (e.g. AR),
easily.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

N/a

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- ~Tests have been written to cover the code changed/added as part of
this pull request~

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: a3b9f90bbc47202868b449590b1c77a1aa70ce92