[US_IX][Workflows] Update SLD in TES to match case compliance (Recidiviz/recidiviz-data#38932)
## Description of the change

In migrating vitals, Dana found some discrepancies between TES and case
compliance outlined
[here](https://docs.google.com/document/d/1VtIseCYU_2EKUzGW6pzSkw1n6myTa9_lklmKnaL5dO4/edit?tab=t.0).

- In order to update TES to more accurately surface clients who are
eligible for a supervision level downgrade we updated TES so that if a
client was awaiting assessment (in the first 45 days of supervision and
no assessment since supervision start) and on `MEDIUM`, this client is
no **ineligible** for SLD. [43 fewer
clients](https://docs.google.com/spreadsheets/d/1uBlfg3qbABXsZYCjciJwhQ2ulbcUA75i8aYO_0TQlZc/edit?gid=1742997253#gid=1742997253)
who are awaiting and assessment and are supervised at `MEDIUM` are no
longer eligible.
- One client is eligible in TES but not case compliance since case
compliance does not count parole board holds as the start of a
supervision super session but TES does and the client is supervised at
`MEDIUM`.
- We also updated that if we have no assessments, we surface clients as
eligible for SLD to `MEDIUM` and this resulted in 3 additionally
eligible clients.
- Lastly, some of the differences were due to the most recent
assessments being from the old system (Level 1-4) and so we added these
levels into the criteria query. This change resulted in 11 more cases
agreeing between case compliance and TES, but four additional clients
getting suggested a downgrade that are NOT suggested a downgrade in case
compliance.

A comparison of the original differences, and the updated results can be
found
[here](https://docs.google.com/spreadsheets/d/1fGpiLBfBFto3X12JvQyPoJ2Iw3Xi-TSGHhb5GwccL-E/edit?gid=770115016#gid=770115016).

**Overall**:
149 eligible in both
17 newly eligible (13 of which have recommended downgrade levels in case
compliance, but 4 of which do not).
43 clients newly ineligible
delta = 26

<img width="732" alt="image"
src="https://github.com/user-attachments/assets/c63baa76-cde5-48e3-a61b-98c0c699d5e7"
/>

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#38336

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e1609c3e70a38209dd519223de66c26f73fac9ec