[Insights] Remove unused excluded officer routes and types (Recidiviz/recidiviz-data#39229)
## Description of the change

Removes unused excluded officer remnants from the insights BE.

In
[dashboards/Recidiviz/recidiviz-data#7313](https://github.com/Recidiviz/recidiviz-dashboards/pull/7313)
we removed the excluded officer remnants from the FE - this PR handles
the corresponding cleanup on the BE. Although this change is just
cleanup, it makes the addition to the SupervisionOfficerEntity type in
Recidiviz/recidiviz-data#39521 easier.

Sentry commented that there's an open issue w `types.py` but it's been
resolved via Recidiviz/recidiviz-data#39158.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#33944

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 8993e9b3fe1847afb92810132e7489e3eeec8b81