[Insights] Add email to SupervisionOfficerEntity type (Recidiviz/recidiviz-data#39521)
## Description of the change

Add officer email to the relevant entity type and querier fns.

We want to display the officer email on the insights officer overview
page but we don't currently send that field to the FE.
Officer email is already present in the `supervision_officer` view and
the officer db schema, so we just need to start querying for it and
sending it to the FE. This PR updates the `SupervisionOfficerEntity`
type to include `email`, modifies the querier where necessary, and
updates relevant tests/snapshots. The next step will be adding email to
the FE officer Zod schema.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

[dashboards/Recidiviz/recidiviz-data#5657](https://github.com/Recidiviz/recidiviz-dashboards/issues/5657)

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 566379e1a88068a47fb32243386b1d41fa02d6f3